+++ src/domains/simple.rs
@@ -196,7 +196,9 @@ mod tests {
         }
 
         fn assert_infer(p: &str, expected: Result<&str, UnifyErr>) {
-            let res = p.parse::<Expr>().unwrap().infer::<SimpleVal>(None, &mut Context::empty(), &mut Default::default());
+            let mut set = ExprSet::empty(Order::ChildFirst, false);
+            let e = set.parse_extend(p).unwrap();
+            let res = set.get(e).infer::<SimpleVal>(&mut Context::empty(), &mut Default::default());
             assert_eq!(res, expected.map(|ty| ty.parse::<Type>().unwrap()));
         }
