+++ src/eval.rs
@@ -17,12 +17,12 @@ pub type Env<D> = Vec<LazyVal<D>>;
 pub enum Val<D: Domain> {
     Dom(D),
     PrimFun(CurriedFn<D>), // function ptr, arity, any args that have been partially filled in
-    LamClosure(Id, Env<D>) // body, captured env
+    LamClosure(Idx, Env<D>) // body, captured env
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub enum LazyValSource<D: Domain> {
-    Lazy(Id, Env<D>),
+    Lazy(Idx, Env<D>),
     Strict(Val<D>),
 }
 
@@ -33,7 +33,7 @@ pub struct LazyVal<D: Domain> {
 }
 
 impl<D: Domain> LazyVal<D> {
-    pub fn new_lazy(child: Id, env: Env<D>) -> Self {
+    pub fn new_lazy(child: Idx, env: Env<D>) -> Self {
         LazyVal {
             val: None,
             source: LazyValSource::Lazy(child, env)
@@ -66,16 +66,16 @@ pub type VError = String;
 
 #[derive(Debug)]
 pub struct Evaluator<'a, D: Domain> {
-    pub expr: &'a Expr,
+    pub expr: Expr<'a>,
     pub data: RefCell<D::Data>,
     pub start_and_timelimit: Option<(Instant, Duration)>,
 }
 
-impl Expr {
-    pub fn eval<D:Domain>(&self, child: Id, env: &mut [LazyVal<D>], timelimit: Option<Duration>) -> VResult<D> {
-        self.as_eval(timelimit).eval_child(child, env)
+impl<'a> Expr<'a> {
+    pub fn eval<D:Domain>(&self, env: &mut [LazyVal<D>], timelimit: Option<Duration>) -> VResult<D> {
+        self.as_eval(timelimit).eval_child(self.idx, env)
     }
-    pub fn as_eval<D:Domain>(&self, timelimit: Option<Duration>) -> Evaluator<D> {
+    pub fn as_eval<D:Domain>(self, timelimit: Option<Duration>) -> Evaluator<'a, D> {
         let start_and_timelimit = timelimit.map(|d| (Instant::now(),d));
         Evaluator {
             expr: self,
@@ -176,35 +176,32 @@ impl<'a, D: Domain> Evaluator<'a,D> {
     }
 
     /// eval a subexpression in an environment
-    pub fn eval_child(&self, child: Id, env: &mut [LazyVal<D>]) -> VResult<D> {
+    pub fn eval_child(&self, child: Idx, env: &mut [LazyVal<D>]) -> VResult<D> {
         if let Some((start_time, duration)) = &self.start_and_timelimit {
             if start_time.elapsed() >= *duration {
                 return Err("Eval Timeout".to_string());
             }
         }
-        let val = match self.expr.nodes[usize::from(child)] {
-            Lambda::Var(i) => {
-                env[i as usize].eval(self)?
+        let val = match self.expr.get_node(child) {
+            Node::Var(i) => {
+                env[*i as usize].eval(self)?
             }
-            Lambda::IVar(_) => {
+            Node::IVar(_) => {
                 panic!("attempting to execute a #i ivar")
             }
-            Lambda::App([f,x]) => {
-                let f_val = self.eval_child(f, env)?;
-                let x_val = LazyVal::new_lazy(x, env.to_vec());
+            Node::App(f,x) => {
+                let f_val = self.eval_child(*f, env)?;
+                let x_val = LazyVal::new_lazy(*x, env.to_vec());
                 self.apply_lazy(&f_val, x_val)?
             }
-            Lambda::Prim(p) => {
-                match D::val_of_prim(p) {
+            Node::Prim(p) => {
+                match D::val_of_prim(*p) {
                     Some(v) => v,
                     None => panic!("Prim `{}` not found",p),
                 }
             }
-            Lambda::Lam([b]) => {
-                Val::LamClosure(b, env.to_vec())
-            }
-            Lambda::Programs(_) => {
-                panic!("todo implement")
+            Node::Lam(b) => {
+                Val::LamClosure(*b, env.to_vec())
             }
         };
         Ok(val)
