+++ src/lib.rs
@@ -9,8 +9,6 @@ mod parse_expr;
 mod types;
 mod util;
 pub mod domains;
-mod alt_expr;
-mod alt_expr_parse;
 
 
 pub use {
@@ -20,5 +18,5 @@ pub use {
     eval::*,
     types::*,
     util::*,
-    alt_expr_parse::*,
+    parse_expr::*,
 };
